+++ CHANGELOG.md
@@ -11,12 +11,17 @@ This project adheres to Semantic Versioning.
 * `graphics::set_letterbox_color` allows you to set the color of the letterbox bars shown in certain scaling modes.
 * Basic support for off-screen rendering/'render to texture' has been implemented, in the form of the `Canvas` object.
 * An `animation_controller` example has been added, showing how to change animations based on the player's input. ([@mgocobachi](https://github.com/mgocobachi) in [#96](https://github.com/17cupsofcoffee/tetra/pull/110))
+* A `from_file_data` constructor has been added to `Font`, for consistency with `Texture` and `Sound`.
 
 ### Changed
 
 * Alpha blending should now work in a more predictable way. This may need further tweaks later on.
 * The renderer now flips drawing automatically when drawing to a framebuffer, due to how OpenGL stores textures. This is similar to how FNA and Love2D handle the same problem.
 * The renderer no longer implicitly re-binds shaders after calling `graphics::present`.
+
+### Deprecated
+
+* `Font::from_data` has been deprecated.
 
 ## [0.2.13] - 2019-03-05